# server/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "PH Regions API",
    "contact": {
      "name": "PH Regions",
      "url": "https://github.com/weaponsforge/ph-regions"
    },
    "license": {
      "name": "MIT"
    },
    "description": "**💡About** \n\n A RESTful API that serves hierarchical location data of the Philippines — including regions, provinces municipalities, and a randomly generated number of barangays per municipality for testing purposes. \n\n **⚠️ Important** <br><br> This API is intended for testing and simulating RESTful API requests from client applications. \n **Note:** The location data may be outdated and does not reflect the most current official records. \n\n **🧩 Alternate Docs** <ul><li>Interactive API documentation (Swagger UI): [http://localhost:3001/api/docs](http://localhost:3001/api/docs)</li></ul>",
    "x-logo": {
      "url": "./assets/images/logo_ph_regions_01.png",
      "altText": "PH Regions logo"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001/",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/provinces/{id}/municipalities": {
      "get": {
        "description": "Get province data by ID including its municipalities",
        "summary": "Get province by ID with municipalities",
        "tags": [
          "Provinces"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Province ID",
              "example": "68bc452bf0a9414a4312e5b1"
            },
            "required": true,
            "description": "Province ID",
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "boolean",
              "nullable": true,
              "default": false,
              "description": "Flag to include the Mongo meta fields (eg., `createdAt`, `updatedAt`) in the response",
              "example": false
            },
            "required": false,
            "description": "Flag to include the Mongo meta fields (eg., `createdAt`, `updatedAt`) in the response",
            "name": "includeMeta",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Object containing one (1) Philippine province data including its municipalities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true,
                      "description": "Server query success flag"
                    },
                    "total": {
                      "type": "number",
                      "description": "Total number of items in the `data[]` array",
                      "example": 1
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "description": {
                          "type": "string",
                          "description": "Summary describing the data set",
                          "example": "Provincial geographic location data of the Philippines"
                        },
                        "source": {
                          "type": "string",
                          "description": "Data source description",
                          "example": "PAGASA Seasonal Rainfall Analysis Table (regions and provinces) and 10-Day Weather Forecast Excel file (provinces and municipalities)"
                        },
                        "dateCreated": {
                          "type": "string",
                          "description": "Date the data set was originally fetched from source(s) in YYYY/MM/DD format",
                          "example": "2022/08/03"
                        }
                      },
                      "required": [
                        "description",
                        "source",
                        "dateCreated"
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "description": "Province ID",
                          "example": "68bc452bf0a9414a4312e5b1"
                        },
                        "regionId": {
                          "type": "string",
                          "description": "Region ID",
                          "example": "68bc452bf0a9414a4312e591"
                        },
                        "name": {
                          "type": "string",
                          "maxLength": 40,
                          "description": "Province name",
                          "example": "Batangas"
                        },
                        "municipalities": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "_id": {
                                "type": "string",
                                "description": "Municipality ID",
                                "example": "68bc452bf0a9414a4312e753"
                              },
                              "regionId": {
                                "type": "string",
                                "description": "Region ID",
                                "example": "68bc452bf0a9414a4312e591"
                              },
                              "provinceId": {
                                "type": "string",
                                "description": "Province ID",
                                "example": "68bc452bf0a9414a4312e5b1"
                              },
                              "name": {
                                "type": "string",
                                "maxLength": 40,
                                "description": "Municipality name",
                                "example": "Agoncillo"
                              },
                              "numDocs": {
                                "type": "number",
                                "description": "Random number",
                                "example": 0
                              }
                            },
                            "required": [
                              "_id",
                              "regionId",
                              "provinceId",
                              "name",
                              "numDocs"
                            ]
                          }
                        }
                      },
                      "required": [
                        "_id",
                        "regionId",
                        "name"
                      ]
                    }
                  },
                  "required": [
                    "total",
                    "metadata",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Query parameter validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": false,
                      "description": "Server query failure flag"
                    },
                    "error": {
                      "type": "string",
                      "description": "Common error message text",
                      "example": "Validation Error"
                    },
                    "message": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "List of error messages",
                      "example": [
                        "Unrecognized key: test"
                      ]
                    },
                    "status": {
                      "type": "number",
                      "description": "HTTP error status code",
                      "example": 400
                    }
                  },
                  "required": [
                    "error",
                    "message",
                    "status"
                  ],
                  "description": "Server error response information"
                }
              }
            }
          }
        }
      }
    },
    "/api/municipalities": {
      "get": {
        "description": "List of municipalities in the Philippines by region or province",
        "summary": "Get municipalities",
        "tags": [
          "Municipalities"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Region ID",
              "example": "68bc452bf0a9414a4312e591"
            },
            "required": false,
            "description": "Region ID",
            "name": "regionId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Province ID",
              "example": "68bc452bf0a9414a4312e5b1"
            },
            "required": false,
            "description": "Province ID",
            "name": "provinceId",
            "in": "query"
          },
          {
            "schema": {
              "type": "boolean",
              "nullable": true,
              "default": false,
              "description": "Flag to include the Mongo meta fields (eg., `createdAt`, `updatedAt`) in the response",
              "example": false
            },
            "required": false,
            "description": "Flag to include the Mongo meta fields (eg., `createdAt`, `updatedAt`) in the response",
            "name": "includeMeta",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Object containing a list of Philippine municipalities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true,
                      "description": "Server query success flag"
                    },
                    "total": {
                      "type": "number",
                      "description": "Total number of items in the `data[]` array",
                      "example": 1
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "description": {
                          "type": "string",
                          "description": "Summary describing the data set",
                          "example": "Municipal geographic location data of the Philippines"
                        },
                        "source": {
                          "type": "string",
                          "description": "Data source description",
                          "example": "PAGASA Seasonal Rainfall Analysis Table (regions and provinces) and 10-Day Weather Forecast Excel file (provinces and municipalities)"
                        },
                        "dateCreated": {
                          "type": "string",
                          "description": "Date the data set was originally fetched from source(s) in YYYY/MM/DD format",
                          "example": "2022/08/03"
                        }
                      },
                      "required": [
                        "description",
                        "source",
                        "dateCreated"
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "description": "Municipality ID",
                            "example": "68bc452bf0a9414a4312e753"
                          },
                          "regionId": {
                            "type": "string",
                            "description": "Region ID",
                            "example": "68bc452bf0a9414a4312e591"
                          },
                          "provinceId": {
                            "type": "string",
                            "description": "Province ID",
                            "example": "68bc452bf0a9414a4312e5b1"
                          },
                          "name": {
                            "type": "string",
                            "maxLength": 40,
                            "description": "Municipality name",
                            "example": "Agoncillo"
                          },
                          "numDocs": {
                            "type": "number",
                            "description": "Random number",
                            "example": 0
                          }
                        },
                        "required": [
                          "_id",
                          "regionId",
                          "provinceId",
                          "name",
                          "numDocs"
                        ]
                      }
                    }
                  },
                  "required": [
                    "total",
                    "metadata",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Query parameter validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": false,
                      "description": "Server query failure flag"
                    },
                    "error": {
                      "type": "string",
                      "description": "Common error message text",
                      "example": "Validation Error"
                    },
                    "message": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "List of error messages",
                      "example": [
                        "Unrecognized key: test"
                      ]
                    },
                    "status": {
                      "type": "number",
                      "description": "HTTP error status code",
                      "example": 400
                    }
                  },
                  "required": [
                    "error",
                    "message",
                    "status"
                  ],
                  "description": "Server error response information"
                }
              }
            }
          }
        }
      }
    },
...
  }
}